requirement.txt
asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
comtypes==1.1.10
cycler==0.10.0
Django==3.1.7
django-paypal==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
et-xmlfile==1.1.0
Flask==1.1.2
idna==3.3
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
legacy==0.1.6
MarkupSafe==1.1.1
matplotlib==3.4.2
mcculw==1.0.0
mysqlclient==2.0.3
newsapi==0.1.1
newsapi-python==0.2.6
numpy==1.20.2
openpyxl==3.0.9
pandas==1.2.3
pendulum==2.1.2
Pillow==8.2.0
PyJWT==2.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PySimpleGUI==4.55.1
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2021.1
pytzdata==2020.1
pywin32==301
requests==2.28.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.11
views==0.3
Werkzeug==1.0.1
xlrd==2.0.1
